Move the edge-lambda tests into a dedicated 'tests' dir
